retrospective-backend/src/main/java/com/groupfour/retrospectivebackend/models/Team.java
package com.groupfour.retrospectivebackend.models;

import org.springframework.data.annotation.Id;
import org.springframework.data.mongodb.core.mapping.Document;

import java.util.ArrayList;

@Document("Team")
public class Team {
    @Id
    private String id;
    private String name;
    private ArrayList<Member> members = new ArrayList<>();

    public Team() {
    }

    public Team(String id, String name, ArrayList<Member> members) {
        this.id = id;
        this.name = name;
        this.members = members;
    }

    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public ArrayList<Member> getMembers() {
        return members;
    }

    public void setMembers(ArrayList<Member> members) {
        this.members = members;
    }
}
